+++ Changelog.md
@@ -11,6 +11,7 @@ Types of changes: New, Improved, Fixed, Removed.
 ## [Unreleased]
 
 - New: Action to remove all trailing whitespace (including whitespace-only lines)
+- New: Action to join lines together with separating spaces
 
 ## [1.1.1] - 30 Jan 2026